requirements_test.txt
pytest-homeassistant-custom-component==0.13.316
ruff>=0.8.0
